+++ src/html/index.js
@@ -12,7 +12,7 @@ const ALL = [
 // We don't use groups of rules such as 'block' and 'inline' for
 // deserialization, because we have a single deserialization rule 'document'.
 //
-// For serialization, there is no has no ambiguity in the Slate
+// For serialization, there is no ambiguity in the Slate
 // format, so we always use all the rules at the same time.
 export default {
     document: [document],