+++ src/markdown/blocks/heading.js
@@ -68,25 +68,41 @@ const deserialize = Deserializer().use([deserializeNormal, deserializeLine]);
 function parseHeadingText(state, level, initialText) {
     let text = initialText;
     reHeading.id.lastIndex = 0;
-    const match = reHeading.id.exec(text);
+    const matchId = reHeading.id.exec(text);
     let data;
 
-    const id = match ? match[2] : null;
-    if (id) {
+    if (matchId) {
         // Remove ID from text
-        text = text.replace(match[0], '').trim();
-        data = { id };
+        text = text.replace(matchId[0], '').trim();
     } else {
         text = text.trim();
     }
 
+    const newState = state
+        .down({ text })
+        .use('inline')
+        .lex();
+
+    // Use the custom ID, or use the id of the last anchor found (see anchors tests)
+    const id =
+        (matchId && matchId[2]) || newState.getProp('lastAnchorId') || null;
+    if (id) {
+        data = { id };
+    }
+
     const node = Block.create({
         type: TYPES[level - 1],
-        nodes: state.use('inline').deserialize(text),
+        nodes: newState.nodes,
         data
     });
 
-    return state.push(node);
+    return (
+        newState
+            .up()
+            // We have consumed any anchor ID that was seen recently
+            .setProp('lastAnchorId', null)
+            .push(node)
+    );
 }
 
 export default { serialize, deserialize };