+++ test/from-markdown/html/anchors-before-heading/output.js
@@ -1,0 +1,15 @@
+/** @jsx h */
+import h from 'h';
+
+export default (
+    <document>
+        <paragraph>
+            This is a simple hack to add anchors to Markdown headings
+        </paragraph>
+        <paragraph>
+            <text />
+        </paragraph>
+        <header_one id="custom-anchor">A heading</header_one>
+        <paragraph>Hey</paragraph>
+    </document>
+);