+++ test/from-markdown/html/inner-text-as-attributes/output.js
@@ -4,10 +4,9 @@ import h from 'h';
 export default (
     <document>
         <paragraph>
-            <text />
-            <html closingTag="</a>" openingTag="<a href=&quot;mylink&quot;>">
+            <link href="mylink" title="">
                 mylink
-            </html>
+            </link>
         </paragraph>
     </document>
 );