+++ test/from-markdown/html/nested-merged/output.js
@@ -4,14 +4,9 @@ import h from 'h';
 export default (
     <document>
         <paragraph>
-            <html closingTag="</cite>" openingTag="<cite>">
-                <html
-                    closingTag="</a>"
-                    openingTag="<a href=&quot;mylink&quot;>"
-                >
-                    link
-                </html>
-            </html>
+            <link href="mylink" title="">
+                link
+            </link>
         </paragraph>
     </document>
 );