+++ test/from-markdown/html/self-closing-tag/output.js
@@ -3,16 +3,6 @@ import h from 'h';
 
 export default (
     <document>
-        <paragraph>
-            <inline
-                type="html"
-                isVoid
-                data={{
-                    closingTag: '',
-                    innerHtml: '',
-                    openingTag: '<br/>'
-                }}
-            />
-        </paragraph>
+        <paragraph>{'\n'}</paragraph>
     </document>
 );