+++ test/from-markdown/html/span/output.js
@@ -3,11 +3,6 @@ import h from 'h';
 
 export default (
     <document>
-        <paragraph>
-            <html closingTag="</span>" openingTag="<span>">
-                Hello
-            </html>{' '}
-            World
-        </paragraph>
+        <paragraph>Hello World</paragraph>
     </document>
 );