+++ test/from-markdown/html/with-alt/output.js
@@ -5,17 +5,7 @@ export default (
     <document>
         <paragraph>
             Some image with attributes{' '}
-            <inline
-                isVoid
-                type="html"
-                data={{
-                    closingTag: '',
-                    innerHtml: '',
-                    openingTag:
-                        '<img src="images/foo.png" alt="foo" width="111" style="margin: 0 auto; display: block">'
-                }}
-            />
-            <text />
+            <image src="images/foo.png" title="foo" />
         </paragraph>
     </document>
 );